requirements.txt
aiohttp==1.3.5
Flask>=0.12.2
gunicorn==19.7.1
peewee>=2.10.2
pyTelegramBotAPI>=3.5.1
requests>=2.18.4
steep-steem>=0.0.13
psycopg2==2.7.3.2

# Use tornado gunicorn workers instead aiohttp
tornado>=4.5.2
